Added icons and changed to promise style
Icons were added to the context menu's in an attempt to make it less
bland.

In addition to interface changes, a new feature is being developed where
you can download a torrent through ruTorrent and it will sync your
remote media to your local server. Requires a 3rd party application.

Changed to promise style callbacks, increasing preformance (not like it
matters), and simplifying the code.